defaultをbox-sizing: border-boxに変更
