+++ frontend/src/entries/Contact.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import styled from "styled-components";
 
-import { media } from "./../commons/StyleUtils";
+import { media } from "./../commons/style";
 
 import * as img from "./../../images/hero.jpg";
